102-fibonacci.c

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -10,14 +10,14 @@
 int main(void)
 {
     int i;
-    unsigned long long int fib1 = 1, fib2 = 2, next;
+    unsigned int fib1 = 1, fib2 = 2, next;
 
-    printf("%llu, %llu", fib1, fib2);
+    printf("%u, %u", fib1, fib2);
 
     for (i = 3; i <= 50; i++)
     {
         next = fib1 + fib2;
-        printf(", %llu", next);
+        printf(", %u", next);
         fib1 = fib2;
         fib2 = next;
     }
